Fixed Minor issues + Changed courses fields for consistency

+++ src/main/java/com/avanse/springboot/DTO/ImageDTO.java
@@ -15,6 +15,6 @@ public class ImageDTO {
 	private int height;
 	private int width;
 	private Date uploadedOn;
-	private Double size;
+	private Long size;
 	
 }
